stack: Split stack.py
The file has grown enough to be splitted into smaller files.
