+++ autopts/ptsprojects/stack/layers/__init__.py
@@ -1,0 +1,36 @@
+#
+# auto-pts - The Bluetooth PTS Automation Framework
+#
+# Copyright (c) 2017, Intel Corporation.
+# Copyright (c) 2024, Codecoup.
+#
+# This program is free software; you can redistribute it and/or modify it
+# under the terms and conditions of the GNU General Public License,
+# version 2, as published by the Free Software Foundation.
+#
+# This program is distributed in the hope it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+# more details.
+#
+
+from .aics import *
+from .ascs import *
+from .bap import *
+from .ccp import *
+from .core import *
+from .gap import *
+from .gatt import *
+from .gattcl import *
+from .gmcs import *
+from .hap import *
+from .ias import *
+from .l2cap import *
+from .mcp import *
+from .mesh import *
+from .micp import *
+from .mics import *
+from .pacs import *
+from .vcp import *
+from .vcs import *
+from .vocs import *
