+++ autopts/ptsprojects/stack/layers/aics.py
@@ -1,0 +1,74 @@
+#
+# auto-pts - The Bluetooth PTS Automation Framework
+#
+# Copyright (c) 2017, Intel Corporation.
+# Copyright (c) 2024, Codecoup.
+#
+# This program is free software; you can redistribute it and/or modify it
+# under the terms and conditions of the GNU General Public License,
+# version 2, as published by the Free Software Foundation.
+#
+# This program is distributed in the hope it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+# more details.
+#
+from autopts.ptsprojects.stack.common import wait_for_queue_event
+from autopts.pybtp import defs
+
+
+class AICS:
+    def __init__(self):
+        self.event_queues = {
+            defs.AICS_STATE_EV: [],
+            defs.AICS_GAIN_SETTING_PROP_EV: [],
+            defs.AICS_INPUT_TYPE_EV: [],
+            defs.AICS_STATUS_EV: [],
+            defs.AICS_DESCRIPTION_EV: [],
+            defs.AICS_PROCEDURE_EV: [],
+        }
+
+    def event_received(self, event_type, event_data_tuple):
+        self.event_queues[event_type].append(event_data_tuple)
+
+    def wait_aics_state_ev(self, addr_type, addr, timeout, remove=True):
+        return wait_for_queue_event(
+            self.event_queues[defs.AICS_STATE_EV],
+            lambda _addr_type, _addr, *_:
+            (addr_type, addr) == (_addr_type, _addr),
+            timeout, remove)
+
+    def wait_aics_gain_setting_prop_ev(self, addr_type, addr, timeout, remove=True):
+        return wait_for_queue_event(
+            self.event_queues[defs.AICS_GAIN_SETTING_PROP_EV],
+            lambda _addr_type, _addr, *_:
+            (addr_type, addr) == (_addr_type, _addr),
+            timeout, remove)
+
+    def wait_aics_input_type_ev(self, addr_type, addr, timeout, remove=True):
+        return wait_for_queue_event(
+            self.event_queues[defs.AICS_INPUT_TYPE_EV],
+            lambda _addr_type, _addr, *_:
+            (addr_type, addr) == (_addr_type, _addr),
+            timeout, remove)
+
+    def wait_aics_status_ev(self, addr_type, addr, timeout, remove=True):
+        return wait_for_queue_event(
+            self.event_queues[defs.AICS_STATUS_EV],
+            lambda _addr_type, _addr, *_:
+            (addr_type, addr) == (_addr_type, _addr),
+            timeout, remove)
+
+    def wait_aics_description_ev(self, addr_type, addr, timeout, remove=True):
+        return wait_for_queue_event(
+            self.event_queues[defs.AICS_DESCRIPTION_EV],
+            lambda _addr_type, _addr, *_:
+            (addr_type, addr) == (_addr_type, _addr),
+            timeout, remove)
+
+    def wait_aics_procedure_ev(self, addr_type, addr, timeout, remove=True):
+        return wait_for_queue_event(
+            self.event_queues[defs.AICS_PROCEDURE_EV],
+            lambda _addr_type, _addr, *_:
+            (addr_type, addr) == (_addr_type, _addr),
+            timeout, remove)
