+++ autopts/ptsprojects/stack/layers/mics.py
@@ -1,0 +1,33 @@
+#
+# auto-pts - The Bluetooth PTS Automation Framework
+#
+# Copyright (c) 2017, Intel Corporation.
+# Copyright (c) 2024, Codecoup.
+#
+# This program is free software; you can redistribute it and/or modify it
+# under the terms and conditions of the GNU General Public License,
+# version 2, as published by the Free Software Foundation.
+#
+# This program is distributed in the hope it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+# more details.
+#
+from autopts.ptsprojects.stack.common import wait_for_queue_event
+from autopts.pybtp import defs
+
+
+class MICS:
+    def __init__(self):
+        self.mute_state = None
+        self.event_queues = {
+            defs.MICS_MUTE_STATE_EV: [],
+        }
+
+    def event_received(self, event_type, event_data_tuple):
+        self.event_queues[event_type].append(event_data_tuple)
+
+    def wait_mute_state_ev(self, timeout, remove=True):
+        return wait_for_queue_event(
+            self.event_queues[defs.MICS_MUTE_STATE_EV],
+            lambda *_: True, timeout, remove)
